# docs/swagger.yaml
basePath: /api/v1
definitions:
  model.AuthForgotPasswordRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  model.AuthLoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  model.AuthLoginResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  model.AuthLogoutRequest:
    properties:
      password:
        type: string
    required:
    - password
    type: object
  model.AuthRenewAccessTokenResponse:
    properties:
      access_token:
        type: string
      expired_at:
        type: string
    type: object
  model.AuthResponse:
    properties:
      auth:
        type: object
    type: object
...
  web.ErrJSONResponse:
    properties:
      error:
        $ref: '#/definitions/web.Error'
        type: object
      process_time:
        type: string
      status:
        type: string
      success:
        example: false
        type: boolean
    type: object
  web.Error:
    properties:
      message:
        type: string
    type: object
  web.JSONResponse:
    properties:
      data:
        type: object
        x-omitempty: true
      process_time:
        type: string
      status:
        type: string
      success:
        example: true
        type: boolean
    type: object
host: localhost:9000
info:
  contact:
    email: support.family-catering@example.com
    name: Family catering API Support
    url: http://www.family-catering.com/support
  description: Documentation for Family Catering API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Family Catering API
  version: "1.0"
paths:
  /auth/forgot-password:
    put:
      consumes:
      - application/json
      description: Auth Forgot password
      parameters:
      - description: logout owner payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.AuthForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Ok
          schema:
            $ref: '#/definitions/web.JSONResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "422":
          description: Unprocessable entity
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
      summary: Auth Forgot password
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login owner using registered email and password
      parameters:
      - description: Create owner payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.AuthLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Ok
          schema:
            allOf:
            - $ref: '#/definitions/web.JSONResponse'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/model.AuthResponse'
                  - properties:
                      auth:
                        $ref: '#/definitions/model.AuthLoginResponse'
                    type: object
              type: object
        "400":
          description: Not Found
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "422":
          description: Unprocessable entity
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
      summary: Login owner
      tags:
      - auth
  /auth/logout:
    delete:
      consumes:
      - application/json
      description: Logout owner
      parameters:
      - description: logout owner payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.AuthLogoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Ok
          schema:
            $ref: '#/definitions/web.JSONResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "422":
          description: Unprocessable entity
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
      summary: Logout owner
      tags:
      - auth
  /auth/renew-access-token:
    get:
      consumes:
      - application/json
      description: Auth Forgot password
      produces:
      - application/json
      responses:
        "200":
          description: Ok
          schema:
            allOf:
            - $ref: '#/definitions/web.JSONResponse'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/model.AuthResponse'
                  - properties:
                      auth:
                        $ref: '#/definitions/model.AuthRenewAccessTokenResponse'
                    type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
      summary: Auth Forgot password
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"